bulds app for production

+++ app/controllers/cars_controller.rb
@@ -1,5 +1,5 @@
 class CarsController < ApplicationController
-    skip_before_action :authorized, only: [:index]
+   
     def index
         cars = Car.all 
         render json: cars
@@ -8,29 +8,36 @@ class CarsController < ApplicationController
 
     def create
         car = Car.create(car_params)
-        byebug
         if car.valid?
             render json: car, status: :created
         else
-            render json: { errors: car.errors.full_messages}, status: :unprocessable_entity
+            render json: { errors: car.errors.full_messages }, status: :unprocessable_entity
         end
     end
 
     def update
         car = Car.find_by(id: params[:id])
         if car
-            car.update(car_params)
-            render json: car
+            if current_user.id == car.user_id
+                car.update(car_params)
+                render json: car
+            else
+                render json: { errors: ["You do not have permissions"]}, status: :unauthorized
+            end 
         else
             render json: {error: "Car not found"}, status: :not_found
-        end       
+        end      
     end
     
     def destroy
         car = Car.find_by(id: params[:id])
         if car
-            car.destroy
-            render :no_content
+            if current_user.id == car.user_id
+                car.destroy
+                render :no_content
+            else
+                render json: { errors: ["You do not have permissions"]}, status: :unauthorized
+            end
         else
             render json: { error: "Car not found"}, status: :not_found
         end
@@ -41,5 +48,7 @@ class CarsController < ApplicationController
     def car_params
         params.permit(:id, :make, :year, :color, :image, :mileage, :price, :dealer_id, :user_id)
     end
+
+   
 
 end
